// Sistemas_Monoliticos/src/modules/product-adm/facade/product-admin.facade.dto.ts
export interface AddProductFacadeInputDto {
  id?: string
  name: string
  description: string
  purchasePrice: number
  stock: number
}

export interface AddProductFacadeOutputDto {
  id: string
  name: string
  description: string
  purchasePrice: number
  stock: number
  createdAt: Date
  updatedAt: Date
}

export interface CheckStockFacadeInputDto {
  productId: string
}

export interface CheckStockFacadeOutputDto {
  productId: string
  stock: number
}